refactor(backend): extract createApi factory + complete PR #13 fixes
PR #13 claimed CSP + Referrer-Policy headers, bcrypt-cap .max(72) on
passwords, .max(254) on email, and date-format regex on
txDate/snapshotDate — none of these were actually present in master.
While refactoring server.ts to enable unit testing, fill the gap.

Refactor: top-level Bun.serve + openDb in server.ts blocked
import-for-test of the HTTP layer. Split into:

- app.ts: pure createApi({db, env}) returning fetch handler
- schemas.ts: zod schemas with the missing .max() + isoDate regex
- helpers.ts: setSecurityHeaders now emits CSP + Referrer-Policy
  + nosniff + frame-deny on every JSON / XLSX / OPTIONS response
- server.ts: thin boot wrapper (env parse, openDb, Bun.serve, static)

Behaviour is otherwise byte-identical: same routes, same error
codes, same DB queries.
